requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
rapidfuzz>=3.6
mutagen>=1.47
google-api-python-client>=2.122
google-auth>=2.29
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
playwright>=1.42
yt-dlp>=2024.4.9
python-dotenv>=1.0
pydantic>=2.6
pydantic-settings>=2.2
